Fix timestamp precision for Rails (#24)
When using the Oj encoder in Rails you may inadvertently not get send sub-second timing on Date and Times.

This update changes the default Oj encoder to be in Rails mode that respects the `ActiveSupport::JSON::Encoding.time_precision` parameter if you are using Rails. If you have already set defaults for Oj those will not be changed.